organize-collection: Special case OK clicks when typing a collection
See the comment in the code for why this is necessary, but in a
nutshell, GtkCellRenderer will emit editing-canceled as soon as any
button is clicked (since we focus out of the editable in order to
click the button), and we don't want "focus-out" to mean "cancel" if
we clicked on the OK button.

Original patch from Cosimo Cecchi for gnome-documents.