gd-main-view: Use item ID and not coords to track button press/release
We only want to use click events where press and release are on the
same item - use the item ID to track this invariant rather than the
event coordinates, since a click sequence is valid even if we moved
over the same element.

Original patch from Cosimo Cecchi for gnome-documents.