embed: Use ClutterImage for preview, and don't switch with GtkNotebook
Instead of using a GtkNotebook lets just use the actors within the
view_actor for switching between overview and preview. Since we are
using a ClutterImage for the preview, using a GtkNotebook would mean
having a GtkClutterEmbed to put the preview actor within it. It is a
good idea to reduce the number of GtkClutterActors and
GtkClutterEmbeds.