view-model: Filter the items from PhotosItemManager based on the mode
It is tempting to use a GtkTreeModelFilter to do the filtering for us.
However there is a problem with that. eg., if the Overview fetched
25 favorites and 25 non-favorites sorted by mtime, and the Favorites
fetched 50, then the Overview would end up showing 75 files instead of
50. Furthermore, when fetching the next batch of files for the Overview
some of them might get sorted before, not after, the ones that had
earlier leaked in from the Favorites. That would be confusing for the
user.