source: Drop the gp:goa-account: prefix from the ID
The prefix is redundant. The goa-account part is unnecessary because
the fact that accounts come from GOA is implied. The gp part is
unnecessary because there is no need for application specific
prefixes. In the end, it is part of GNOME and not a particular
application.

This simplifies addition and removal of GoaObjects in
PhotosSourceManager.