Show the contents of a collection from collections and favorites modes
Selecting a collection inside the collection or favorites mode should
display the items inside the collection.

So we modify the queries returned by the PhotosTrackerController and
PhotosOffsetController implementations so that they don't use an query
flags when an active collection is present. Similarly, we don't remove
anything from PhotosViewModel when they stop being a collection or
favorite when a collection is active