base-item: Refactor thumbnail creation for remote items
A new pure virtual method called "create_thumbnail" has been added. It
is meant to be run in a thread and implementations should take care of
creating the thumbnail image file according to the Thumbnail Managing
Standard [1].

Currently GnomeDesktopThumbnailFactory is used for creating local
thumbnails. Thumbnailing Flickr content is yet to be implemented.

[1] http://specifications.freedesktop.org/thumbnail-spec/thumbnail-spec-latest.html

Fixes: https://bugzilla.gnome.org/697675