base-item: Refactor opening in external application
We have been using gtk_show_uri to open the URI till now. However this
doesn't work for Flickr because the URI points to a JPEG file, and not
the Flickr page for the particular photo. A virtual method with a
default implementation was added so that we can tweak the behaviour
for Flickr.