header-bar: Fix regressions after move to GtkHeaderBar
Unlike its Gd counterpart, GtkHeaderBar does not consider the
custom_title widget as an internal child. Hence it is exposed in
gtk_container_get_children, and is affected by gtk_widget_show_all
being applied to the header bar.

Ensure that the stack_switcher is not destroyed when the header bar
is cleared. Also, it should not be affected by show_all, otherwise
buttons corresponding to private children might show up.