Add PhotosSearchContext and make PhotosApplication implement it
We want to split the application into two search contexts - one is for
searches initiated from within the application, and the other is for
searches initiated from gnome-shell.

For that to happen certain classes which were singleton, and hence
global to the entire application, now has to have a separate instance
for each context. Each context will, therefore, have a state that
encapsulates the instances of these classes that correspond to it.

Currently, PhotosApplication is the only context. The future
PhotosShellSearchProvider will be another implementation.