base-manager: Let the PhotosQueryFlags be passed to get_where
We want to have a separate mode for showing the search results. We do
not want to filter the existing modes (ie. overview, collections and
favorites) during search.

Therefore, the where clause should only be dynamic (based on the
current active object) in the search mode. Otherwise it should be
static. Having the access to the PhotosQueryFlags lets us do that.