base-manager: Make get_filter a virtual method
... and make PhotosSearchTypeManager and PhotosSourceManager override
it. This is required for implementing a separate mode for showing the
search results. The filter should only be dynamic (based on the
current active object) in the search mode. Otherwise it should be
static.

Being able to tweak the filter based on the PhotosQueryFlags from the
individual managers lets do that.