collection-icon-watcher: Be careful with the lifetime of the collection
... because it is not owned by the watcher. Instead the watcher is
owned by it. We keep a reference to the watcher while doing various
asynchronous operations, but that does not mean that the collection is
still valid. It might have been destroyed while we were waiting for the
asynchronous operations to complete.

Therefore, add a weak reference to the collection and check its
validity wherever applicable.