Shed redundant SPARQL queries and PhotosViewModel updates
We were refreshing all the PhotosTrackerControllers and clearing
PhotosItemManager multiple times when the active collection changed.
This is too much. Only those components that are tied to the current
mode should react.

This means we should only refresh the PhotosTrackerController
corresponding to the current mode, and instead of clearing the whole
manager we should only clear the relevant PhotosViewModel.

Since we are not clearing PhotosItemManager on every refresh, we will
miss some "object-added" signals because the items found while
refreshing might already be part of the manager. We need to adjust for
that.

We will also need to be more strict when adding items to the model.
Particularly when there is an active collection in one of the modes.