mode-controller: Add photos_mode_controller_go_back
We have to track the previous or old mode in a number of places. Either
to implement the back button or escape key or to figure out which view
to switch to when the search constraints are removed. This helps with
that.

Secondly, we need to use a stack instead of a single variable for this.
A single variable is not enough to track transitions like:
 - launch application in overview
 - search for something (ie. overview -> search)
 - preview a search result (ie. search -> preview)
 - go back to search results (ie. preview -> search)
 - remove search constraints (ie. search -> ???)

At this point, with a single variable, preview is the "old mode", not
the overview. This is where having the whole history in a stack is
useful.

Fixes: https://bugzilla.gnome.org/725329