Skip empty collections
Skipping empty collections in the collection view is easy. Adding
'?item a nie:InformationElement; nie:isPartOf ?urn.' to the WHERE
clause of the SearchType is enough for that.

However, it is not so easy when we want to have both non-empty
collections and items. eg., in the search view. In that case we need
to use a 'OPTIONAL {?item a nie:InformationElement; nie:isPartOf ?urn}'
and 'COUNT (?item)', and nested SELECTs to filter out empty
collections while still catching nmm:Photo instances. We use the
same FILTER as the favorites view instead of the generic catch all one
to avoid bloating the query.

Fixes: https://bugzilla.gnome.org/725589