view-container: Prevent the icons from shifting around
The album and photo icons are not exactly of the same size because we
do not frame the albums. This causes the icons to shift horizontally
when moving from a view with only albums to one that has both albums
and photos. So lets set an item width on the GtkIconView to avoid this.

Fixes: https://bugzilla.gnome.org/698941