base-item: Clean up the pixbuf loading code
Rename pristine_icon as original_icon and use a single function for
setting and updating the effects. Be careful not to change the state
of the item if we failed to generate original_icon.

Streamline the way original_icon (formerly pristine_icon) is set. Set
it directly, instead of passing it via icon. Use icon only to
represent the final thumbnail with the rendered emblems and a frame.

Original patch from Cosimo Cecchi for gnome-documents.

Fixes: https://bugzilla.gnome.org/726903