Drop GdHeaderMenuButton
A GdHeaderMenuButton is a GtkMenuButton that either has a 16px icon,
tooltop text and the image-button style class; or a label and the
text-button style class.

GtkButton does not offer a use-markup property so we need to set it
directly on the child GtkLabel. We also need to guard against NULL
strings because gtk_button_set_label is not marked as 'allow-none'.

Fixes: https://bugzilla.gnome.org/733850