application: Don't crash if a primary instance is already present
The dbus_unregister method can be invoked more than once and be
interleaved with dispose. We should not try to unexport a NULL
SearchProvider. Let's clear the reference and check for NULL in
dbus_unregister, and do the same in dbus_register if we failed to
export.

Based on code from gnome-terminal.