main: Locale fix
Ensure that each part of the locale is set according to the current
environment variables. Otherwise we leave ourselves open to locale
mismatches between the Tracker database and the application.

This can lead to the DB's locale changing on start; or failure to use
Tracker's 'direct' backend and falling back to 'bus'.

We were quite reliably hitting this problem when running the test
suite.