Use a fallback name in the header bar when there is no nie:title
Add a pure virtual method to PhotosBaseItem so that different
implementations can offer their own fallbacks. For local content we
use the filename without the extension; for remote content we use
something like "Facebook — 2nd January 2013", where the second part is
the mtime.

This is not implemented, yet, for media server items.

https://bugzilla.gnome.org/show_bug.cgi?id=733214