application: Create the miner proxies asynchronously & only when needed
We only need them when an application window is shown, so there is no
point in creating the miner proxies during startup which is shared
between the search provider and the application. So, let's tie them to
the lifetime of the window.

Plus, creating them asynchronously is the right thing to do, and we
can't sneak in G_DBUS_PROXY_FLAGS_DO_NOT_LOAD_PROPERTIES because we use
the properties.

https://bugzilla.gnome.org/show_bug.cgi?id=744627