preview-nav-buttons: Re-evaluate the ability to navigate at each step
We only call photos_preview_nav_buttons_show when we first move to the
preview, instead of every time a GeglNode is set. This is logical since
we don't want to affect the visibility of the buttons as we navigate.
However, we need to re-evaluate the ability to navigate in each
direction after every step. Otherwise the controls won't match the
realities of our location within the model.

There was an earlier broken attempt to address this in
113d48bc7ea6a3cc7e40bd2fec1e6a85684e942f and it became worse after
949246eab98663c8d84d575cbe785f982f4c5de0

https://bugzilla.gnome.org/show_bug.cgi?id=746152