preview-nav-buttons: Untie keynav from the visibility of the buttons
We can not tie the ability to go forward or backward to the visibility
of the buttons. Otherwise, if we have hidden them due to an internal
condition (eg., no mouse movement or tapping on the touchscreen), then
the ability to navigate using the keyboard will also be disabled.

Fall out from b3ed6640d3cd2ef9052b9fa18e515b1c9c515380

https://bugzilla.gnome.org/show_bug.cgi?id=746152