search-match: Silence -Wformat-nonliteral
Can't think of a better way than using a #pragma. The filter is the
template used to generate the SPARQL blurb, and term is the text input
by the user. We can't easily turn filter into a string literal because
we have to create multiple SearchMatch instances with different filter
templates. We can't construct a va_list either because our function is
not variadic.