google: Look at the cache before hitting the network
This makes sense because network and server failures should not fail
the operation if we have an entry in the cache.

This changes the cache structure. Instead of using
GDataEntry:content-uri as the file name, we use a hash of the
nao:identifier. Hashing is necessary because the nao:identifier is not
a valid file name.