gegl-gtk-view-helper: Don't redraw on every GeglNode::computed
GeglNode::computed is emitted after each chunk is processed. Drawing it
chunk by chunk would create a visible checkerboard-like effect. To
avoid it, PhotosPreviewView queues a draw after the entire GeglNode
has been processed, so that the entire image can be drawn in one shot.

There was a bug in GEGL which prevented GeglNode::computed from being
emitted for graph nodes, so this was merely code that was not being
used. Now that the bug has been fixed, we need to remove this code to
prevent the checkerboard.

https://bugzilla.gnome.org/show_bug.cgi?id=750938