gegl-gtk-view-helper, preview-view: Merge AUTOSCALE_CONTENT logic
We want AUTOSCALE_CONTENT to always centre the image inside the
widget's allocation irrespective of the current origin. GeglGtkView
wasn't doing this. So we had our own code to centre the image by
calculating a new origin and setting it, under the following scenarios:
  - Setting a new GeglNode
  - Completion of graph processing (ie. the last GeglNode::compute)
  - Size allocation has changed (ie. GtkWidget::size_allocate)

Calculating a new origin also required us to calculate a new scaling
factor, just like GeglGtkView. Unlike our code, GeglGtkView was doing
it 'iteratively'. It would convert the image's bounding box to widget
co-ordinates by using the existing scaling factor, and then adjust it
by multiplying with a factor. Setting the new scale and the origin
would (needlessly) re-run the AUTOSCALE_CONTENT logic until it reached
a steady state. Ideally, the subsequent runs of the logic would be
NOPs, but inaccuracies of floating arithmetic often introduced small
errors.

Now, we completely ignore the existing scaling factor, and calculate
the new scale and origin from scratch. We also don't re-run the logic
after setting the new values.

https://bugzilla.gnome.org/show_bug.cgi?id=747446