spinner-box: Show it only when revealed (and running)
We need to hide the GtkRevealer when the child is not revealed to
ensure that the GdkWindow also goes away. Otherwise, it stays overlaid
and interferes with the crop tool. For example, the crop tool changes
the GdkCursor depending on where the mouse is relative to the crop
area. If the SpinnerBox's GdkWindow is in the overlay then the mouse
pointer changes while passing over the invisible spinner.