Let users page through search results in the preview
We were losing the search results when turning off the searchbar before
moving to the preview. We avoid this by freezing the TrackerControllers
so that they don't react to any changes in the search parameters.

The TrackerControllers are thawed *after* we are back and the search
state has been restored to avoid needless queries.

https://bugzilla.gnome.org/show_bug.cgi?id=741593