searchbar: Change the GAction APIs that we use for consistency
Settling on a smaller subset of the GAction, GActionMap and
GActionGroup APIs makes it easier to grep them. Otherwise, one has to
grep for a bunch of different strings to find out all the uses of the
actions.

Remember, that since we are connecting to GSimpleAction::change-state,
we have to take care of setting the state ourselves.