embed: Remove a wrong comment
The active item and the active collection being identical doesn't
imply that the active item is NULL (or that we are in the overview).
They can be identical even when we are viewing a collection. The code
is correct, but the comment was wrong.

Fall out from 5f910e73e58b50f791c48ceae9a9d63ab8ee86a6

https://bugzilla.gnome.org/show_bug.cgi?id=741593