application: Set priority rdf:types only when a window is created
The startup is shared between the application and the search provider.
There is no need to set the priority for the search provider because we
want to use whatever is already there in tracker's database. Moreoever,
searching in gnome-shell spawns a bunch of search provider processes,
and having all of them set their own priorities will be ineffective.

We generalized the GCancellable driving the creation of the miner
proxies to drive all asynchronous operations happening at window
creation.

https://bugzilla.gnome.org/show_bug.cgi?id=744627