main-toolbar: The back button should always be the left-most element
The back button would show up after the remote display button if we
moved to a collection while the remote display manager was active. Fix
it by updating the remote display button (ie. destroying and re-adding
it) after we have updated the back button on active-collection-changed.
This matches what we were doing for selection-changed and other mode
changes.

https://bugzilla.gnome.org/show_bug.cgi?id=732799