Use GActions to represent the next and previous operations
We were exposing too many internal details of classes (eg.,
photos_embed_get_preview) to hook up the key press event in the
MainWindow to the actual implementation in PreviewNavButtons. Plus we
had our own set of signals to propagate the event to the rest of the
code (eg., PreviewView). We can avoid all this by using GActions. It is
simply the right thing to do.

https://bugzilla.gnome.org/show_bug.cgi?id=733660