view-model: Don't keep it alive just for the timeout
The whole point of updating the OffsetController's count is to ensure
that we can load more items and show the "No Results" page. If the
ViewModel is about to be destroyed, then the ViewContainer is also
gone, and the count is useless without a view. So we might as well
spare ourselves a needless query.

https://bugzilla.gnome.org/show_bug.cgi?id=757199