selection-toolbar: Avoid O(n) logic
To check whether a GList is non-empty, it is faster to check it against
NULL. We can similarly check if there is only one element by looking at
the pointers. In those cases where we really do need the length, we can
calculate it ourselves if we are already iterating over the list.

https://bugzilla.gnome.org/show_bug.cgi?id=757363