gegl-gtk-view, gegl-gtk-view-helper: Fix HiDpi rendering
Unlike Cairo and GTK+, GEGL is unaware of the scale factor and only
deals in device pixels. We should take the scale factor into account
so that we can exploit the higher DPI to render sharper images by
pumping in more pixels.

Since the GdkWindow that we are drawing on has a scale factor, we
set up our Cairo image surface to have the same. Otherwise, the pixels
will be upsampled when drawing to the window.

We use the allocation in device pixels when calculating the scaling
factor so that GEGL can pump in more pixels to make it look sharper.
Same for calculating the co-ordinates of the rectangle to render from
the GeglNode.

We use the widget's allocation in device pixels to decide whether an
image is big enough to be downscaled. Therefore, if it is small enough
to fit on a HiDpi display, but not on LoDpi, it is not considered as
'big'. It isn't scaled, and will look sharper and smaller on HiDpi as
compared to LoDpi. The other option would be to upscale it to take up
the extra pixels, but I think a sharper render is a better default.
Plus we don't want to optimize for tiny images where the smaller size
would make them unrecognizable. We can let users adjust the zoom if it
proves to be an issue.

This makes HiDpi screenshots render at the same sharpness and size as
the real widgets on a similar display. However, LoDpi screenshots will
look smaller on HiDpi, just as HiDpi screenshots will look larger on
LoDpi. Since it is more common to view screenshots taken on the same
computer, this is an acceptable trade-off for the time being.

https://bugzilla.gnome.org/show_bug.cgi?id=747446