application: Use the right mechanism to stay alive across async calls
Normally, the application will be shutdown and the process will die
once the top level window is destroyed and all the ready sources have
been processed. Holding a normal reference will not keep it alive.
The way to do that is to increase the use count by calling
g_application_hold.

This is not a problem, other than the fact that we don't get a chance
to cleanly free all resources.