preview-view: Don't put everything in a single GtkScrolledWindow
There was never any reason to derive from GtkScrolledWindow and put
everything in a single scrolled window. That was a stupid move on my
part. Instead each view and the palette should have separate scrolled
windows.

This is not a big issue at the moment, but it will be when the palette
grows further or we implement zoom.