gegl-gtk-view, gegl-gtk-view-helper, tool-crop: Rename scale to zoom
Terms like "device_scale", "scale" or "scale_factor" are widely used
in Cairo and GTK+ to refer to the HiDpi scale factor that maps from
logical window co-ordinates to the actual device pixels. It's
confusing to use the same term to refer to the item's zoom
factor. Especially when both concepts are used in the same class.

I don't know if we should rename "autoscale", and, if we do, what the
new name should be. We can revisit it later.

https://bugzilla.gnome.org/show_bug.cgi?id=758352