gegl-gtk-view-helper, tool-crop: Fix cropping on HiDpi
Just like other higher level Cairo and GTK+ API, GeglGtkView's public
API should only use logical window co-ordinates, and the scaling logic
for sharper rendering on HiDpi should remain internal to the widget.
On the other hand, GEGL is unaware of the scale factor and only deals
in device pixels. So, whenever we touch the pipeline we should
continue using the device pixels.

We simplify the arithmetic a bit (see below) by continuing to use
device pixels in photos_base_item_get_bbox_source, but keeping the
extents of the zoomed source in logical co-ordinates.

This means that ToolCrop, which is mostly drawing and event handling
code, doesn't need to worry about the scale factor. In theory, we
would have to convert from logical co-ordinates to device pixels when
interfacing with GEGL, but, in practice, things line up nicely that we
don't have to.

https://bugzilla.gnome.org/show_bug.cgi?id=758352