base-item, local-item: Track both the default GAppInfo and it's name
... for LocalItem. The default GAppInfo can change behind our backs and
we don't want the actual application and the label shown in the UI to
get out of sync.

If we have a default GAppInfo, then explicitly use it to launch the
application instead of just relying on gtk_show_uri. This defends us
against broken defaults.

Original patch from Bastien Nocera for gnome-documents.

https://bugzilla.gnome.org/show_bug.cgi?id=740434