view-model: Break a reference cycle
A ViewModel instance can have "hidden" references held by the
GtkTreeRowReferences of each item that is present in the model. Since
items are owned by the ItemManager, this means that ItemManager can
hold a strong reference to a ViewModel. Therefore, ViewModel itself
shouldn't hold a strong reference to ItemManager. Otherwise, we would
have a reference cycle.