application, base-item: Rework photos_base_item_save_async
The UX design expects the user to specify the sub-directory within
which the item will be saved. Plus, we can simplify things a bit by
using the BaseItem's filename, which must be present if it is
editable, instead of trying to generate one.

Therefore, we rework the API to accept the sub-directory and not the
complete path to the final image.

https://bugzilla.gnome.org/show_bug.cgi?id=759363