application, base-item: Conditionally enable editing
Only local items can be edited. We require that the item has finished
loading and we are in the preview.

There are some high level challenges and questions involved in editing
remote items - do we upload the edited version; do we replace the
remote source with the edited version; etc.. So, let's keep it simple
for the time being.