utils: Add photos_utils_create_buffer_from_node
In the following patches, we need a way to get the output of a fully
processed GeglNode as a GeglBuffer, which we are going to use in a
separate thread to run various throwaway operations. eg., guessing the
number of bytes if the buffer is JPEG compressed, or zooming it before
saving.