base-item: Save everything other than PNGs as JPEGs
This is a reasonable strategy to start with. It let's us avoid writing
size guessing code for every possible image format out there. Apart
from RAW images, PNGs and JPEGs are the most common formats out there,
and JPEG is a good option for exporting RAWs.

We can add support for more formats as we go along.