application: Sanity check GEGL operations before creating the window
It's too easy to miss some GEGL operation due to not having the
relevant dependencies while building. Let's check these things upfront
to avoid weird behaviour later on.

We perform these checks when creating the main window because the
earlier stages are shared with the search provider. The search
provider doesn't use GEGL and we want to keep it as lean as possible.