Drop PHOTOS_OPERATION_INSTA_PRESET_GRAY
Gray was mainly a placeholder because we didn't have an even number of
magic filters. Now we have Hefe. Instagram Inkwell is considered to be
the same as Gray, but it can also be achieved using the saturation
tool. There is no need to have a magic filter for it.