Rewrite gegl-gtk as PhotosImageView
Our copy of gegl-gtk had diverged quite a bit:
 - We don't use a processor inside the widget.
 - Our AUTOSCALE_CONTENT logic always centres the image.
 - Sharper renders on HiDpi displays.
 - We don't care about GTK+ 2.x or the absence of Cairo.

The widget / helper split is also largely useless for us. While it
will be good to have tests for the automatic centring and scaling code,
we don't need a separate helper class for it.

While some of our changes can be upstreamed, we are ultimately going
to use the new image viewer widget in GTK+. We are also going to
use transitions when the output of the GeglNode changes. Having only
the bits that we need in a single class, and getting rid of the extra
fluff will help us evolve this code in future.

Since PhotosImageView largely constitutes code written by me, we can
license it as GPLv2+, unlike gegl-gtk, which was LGPLv3+.

https://bugzilla.gnome.org/show_bug.cgi?id=761559