application, preview-view: Move the "edit-done" implementation
In the following patch, we want to ensure that the current tool is
deactivated before we leave the edit mode. Otherwise, tools like
ToolCrop won't get a chance to apply their changes.

To do this we need to access objects that are local to PreviewView and
are not visible to the Application. Therefore, we need to move the
implementation of the "edit-done" GAction to a more suitable spot.

https://bugzilla.gnome.org/show_bug.cgi?id=760931