Handle asynchronous activation of PhotosTool instances
Some tools can be activated asynchronously. eg., PhotosToolCrop. We can
not queue a draw until the entire activation step has finished.
Otherwise, we will be asking the tool to draw while it is in an
incoherent state. Therefore, add a new signal to let PhotosTool
sub-classes indicate that they have finished activating.