tool-crop: Fix reset
When resetting a crop, nothing was actually done regarding the
gegl:crop operation still in the pipeline. Because of that, the crop
area would go out of bounds when the user tried to crop the item
again, making it impossible to crop and/or resize the area.

Therefore, instead of adjusting the gegl:crop to cover the entire item
when the tool is active, we remove it from the pipeline altogether.
We add it back, as usual, if the user ended up cropping.

https://bugzilla.gnome.org/show_bug.cgi?id=761683