pipeline: Avoid using photos_pipeline_undo
The undo API, the way it is implemented, doesn't match with the current
realities of our UX. Therefore, we need to replace it with something
that meets our needs.

In this case, we were using it to ensure that the graph was empty
before resetting it. Instead we can add a pre-condition that asserts
the emptiness of the graph.