application, pipeline: Add API to revert a series of edits and use it
The undo operation doesn't fit with the realities of our current UX
because we don't allow undoing a single edit. We always revert a series
of edits. In the following patches we will implement snapshots which
will let us revert a specific set of edits.

Therefore, a revert operation to remove all the operation nodes in the
GEGL graph is clearer and more useful.

https://bugzilla.gnome.org/show_bug.cgi?id=761683