pipeline, utils: Refactor removal of children into a separate function
We can't use gegl_node_remove_children because it would also remove the
proxies, which we want to avoid. We are going to use this when creating
the graph from XML because we only want to re-create the internals of
the graph and not the whole graph itself. Otherwise, external code
holding references to the graph and its proxies will get confused when
we revert the pipeline to an earlier snapshot.

https://bugzilla.gnome.org/show_bug.cgi?id=762046