pipeline: Also look at the compat-name when deserializing the graph
Sometimes, we might want to use a different name for an operation in
the XML serialization of the graph. Since the XML is created with the
GeglOperation's "name" key, we would use "compat-name" for the
internal name of the operation. Therefore, we need to look at the
"compat-name" when creating the mapping from operation names to
GeglNodes.