Add a GAction to redraw the current PhotosImageView
This will be useful to queue redraws of the current PhotosImageView
from other parts of the application. eg., in future, we will add
PhotosDoneNotification that will allow undoing the effect of clicking
"done" in the edit mode. This will need to queue a redraw after
reverting the pipeline to its previous snapshot.

https://bugzilla.gnome.org/show_bug.cgi?id=763096