Fix the reference counting for NotificationManager
We were not finalizing NotificationManager when the main window is
destroyed. Every call to photos_notification_manager_dup_singleton,
except the first one, invoked both g_object_ref and g_object_ref_sink.
This is wrong because it leaks a reference.

NotificationManager is odd in the sense that it is a GInitiallyUnowned
(because it is a GtkWidget) and a singleton at the same time. Let's
ensure that notification_manager_dup_singleton always sinks the
floating reference. This hides all the subtleties of floating
references from the clients.

https://bugzilla.gnome.org/show_bug.cgi?id=763908