pipeline: Mark nodes as passthrough instead of actually removing them
The strategy of removing a GeglNode from the graph when it is no
longer needed doesn't match with our desire to have pre-determined
positions for certain operations. It will be easier if we can mark
certain nodes as unused without removing or changing their position
in the graph. In fact, right now, we ensure that photos:insta-filter is
the first operation in the pipeline by creating it before any other
operation is added, and hide its existence by setting it to its NONE
preset. When photos:insta-filter is actually asked for, we update the
properties of this node. We want to extend this technique to cover
other nodes as well.

Unfortunately, simply adding a GeglNode and setting it to its no-op
values is not ideal. It will confuse photos_pipeline_get. It won't be
able to distinguish between no-op values actually set by the user, and
those that were set internally to mimic a removal. To be fair, it is
not clear whether that matters, though. More importantly, tracking the
no-op values can be tricky in practice. We can assume the default
values of the properties represent the no-op state. However, a quick
look at the generated XML won't tell us which nodes are no-ops, unless
we remember the set values, which is hard.

One option is to set GeglNode:name to gegl:nop when we want to remove
it. This is almost perfect - we can check the name in
photos_pipeline_get to know if an operation is meant to be absent, and
they will stand out in the XML. However, this won't work because
gegl:nop nodes are skipped during serialization. This means that when
the graph is deserialized, the gegl:nop nodes will be missing and it
will defeat our whole strategy.

Instead, we can mark them as passthrough nodes. They have the same
effect as gegl:nop with the benefit that they are not skipped during
serialization.

https://bugzilla.gnome.org/show_bug.cgi?id=764801