pipeline: Assign a fixed position to gegl:crop
We need to ensure that gegl:crop is placed before any spatial
operation. eg., an insta-filter preset that uses a vignette. Otherwise
the gradient will get clipped.

Also, having it early in the graph has the benefit of reducing the
number of pixels that successive operations have to deal with.

https://bugzilla.gnome.org/show_bug.cgi?id=764801