base-item: Disable favorite albums when reading from the database
If the member variable doesn't reflect the final outcome, then we will
have to call photos_base_item_is_favorite instead of directly checking
the variable inside the class. eg., when deciding whether or not to
show the emblem. This seems a bit wasteful, when we don't care about
the actual value in the database.

https://bugzilla.gnome.org/show_bug.cgi?id=764680