application: Add API to prevent the main window from getting destroyed
We don't want the application to shut down while certain critical
asynchronous operations are in flight. eg., saving the pipeline or
deleting an item.

The problem with g_application_hold is that it doesn't prevent the
main window from getting destroyed, and some of the callbacks of
these operations expect the widget hierarchy to be in place. While we
can try to make things more robust on a case by case basis, it will be
hard to ensure that all corner cases have been covered.

It is much safer to temporarily override the default
GtkWidget::delete-event handler to only hide the PhotosMainWindow
instead of destroying it.

https://bugzilla.gnome.org/show_bug.cgi?id=764076