item-manager: Fix the lifetime of the changes-pending handler
TrackerChangeMonitor is a singleton which can outlive the ItemManager.
Therefore, we should use g_signal_connect_object to ensure that the
TrackerChangeMonitor::changes-pending handler is disconnected when
ItemManager is destroyed.

Note that right now ItemManager is the only user of
TrackerChangeMonitor. So this shouldn't have been a problem. However,
TrackerChangeMonitor doesn't cancel its internal asynchronous
operations and sources during destruction. It takes a reference on
itself to stay alive and can thus outlive the ItemManager.

https://bugzilla.gnome.org/show_bug.cgi?id=767892