single-item-job: Remove redundant function call
Destroying a TrackerSparqlCursor is enough to close it. No need to
explicitly close it.

In the following patch, when we convert SingleItemJob to use a
GAsyncResult based asynchronous API, we can't close the cursor like
this. The GAsyncReadyCallback will be called in the next iteration of
the GMainLoop, and strictly speaking we shouldn't call
tracker_sparql_cursor_get* on a closed cursor. It works today, but
might not work in the future.

https://bugzilla.gnome.org/show_bug.cgi?id=764086