base-item: Simplify photos_base_item_can_edit
Unless the Tracker metadata is corrupted, every BaseItem will now have
a valid filename. Therefore, there is no point in using it to determine
if a BaseItem can be edited or not. In case the metadata is corrupted,
it should be handled by the relevant g_return_if_fail* checks.

https://bugzilla.gnome.org/show_bug.cgi?id=766162