base-item, export-dialog: Calculate the size & zoom in guess_save_sizes
By making photos_guess_save_sizes_finish return a more descriptive
data, we avoid having the caller deal with a lot of detail. Like
checking if downscaling is needed and by how much, and interpolating
the dimensions based on it.

This will be useful for adding support for exporting albums and
multiple images.

https://bugzilla.gnome.org/show_bug.cgi?id=760839