view-container: Consolidate searchbar and mouse activation
ViewContainer->current_path is only set in response to
GdMainView::item-activated, which is emitted due to a mouse click.
When we activate by hitting ENTER in the searchbar, current_path might
not have been set (or set to something that is not the first item).
Therefore, when we try to use it to set the model for PreviewView, it
leads to:
  CRITICAL **: photos_preview_view_set_model: assertion
    'current_path != NULL' failed

Let's handle searchbar activation by triggering the same code path
that is used for mouse activation.

https://bugzilla.gnome.org/show_bug.cgi?id=768130