base-item: Simplify tracking the nie:DataSource
Usually we will have a large number of BaseItems pointing to a small
number of nie:DataSources. The URNs of these nie:DataSources look like:
 - gd:goa-account:account_1442355024_0
 - urn:nepomuk:datasource:d289a99b2854ad9e9d19d497c112f6d9

This means that there will be several copies of a handful of strings.
Memory is fragmented and wasted.

Most of the time we want the online Source of the BaseItem, if any,
and not the nie:DataSource's URN. It is unnecessary work to go via the
SourceManager to resolve the URN.

Therefore, let's just keep a reference to the Source if the URN points
to an online nie:DataSource, and a GQuark to track the URN itself. That
means we replace a pointer and heap allocated string with two pointers
and a 32-bit integer. Some sub-classes were using one of those two
pointers anyway. We can't say how much memory or fragmentation we save
without actual measurements, but it isn't any worse and it does
simplify the rest of the code. The little extra work while reading the
TrackerSparqlCursor data will be masked by the disk I/O and IPC
involved.

https://bugzilla.gnome.org/show_bug.cgi?id=768766